[SPARK-57885][SQL] Promote `EmptyRDDWithPartitions` to a top-level class
### What changes were proposed in this pull request?

This PR promotes `EmptyRDDWithPartitions` from the `CoalesceExec` companion object to a top-level class in `org.apache.spark.sql.execution`, and uses it in the other places that hand-rolled an empty single-partition RDD via `sparkContext.parallelize(...)`.

Note that `EmptyRDDWithPartitions extends RDD[InternalRow]`. So, this is under `sql` module unlike `EmptyRDD` in `core` module.

### Why are the changes needed?

To deduplicate the "empty RDD with one partition" idiom with a single purpose-named class consistently.

```scala
- sparkSession.sparkContext.parallelize(Array.empty[InternalRow].toImmutableArraySeq, 1)
+ new EmptyRDDWithPartitions(sparkSession.sparkContext, 1)
```

https://github.com/apache/spark/blob/17e7e1479eadcdc3a97e06e3b69127576df17412/core/src/main/scala/org/apache/spark/rdd/EmptyRDD.scala#L28

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

Generated-by: Claude Fable 5

Closes #56967 from dongjoon-hyun/SPARK-57885.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
