+++ sql/core/src/main/scala/org/apache/spark/sql/execution/EmptyRDDWithPartitions.scala
@@ -1,0 +1,42 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.sql.execution
+
+import org.apache.spark.{Partition, SparkContext, TaskContext}
+import org.apache.spark.rdd.RDD
+import org.apache.spark.sql.catalyst.InternalRow
+
+/**
+ * A simple RDD with no data, but with the given number of partitions. Used by operators that
+ * must produce at least one (empty) partition even when their input has none, e.g. to honor a
+ * reported `SinglePartition` output partitioning, or to set up a single write task that writes
+ * the metadata for an empty write.
+ */
+class EmptyRDDWithPartitions(
+    @transient private val sc: SparkContext,
+    numPartitions: Int) extends RDD[InternalRow](sc, Nil) {
+
+  override def getPartitions: Array[Partition] =
+    Array.tabulate(numPartitions)(i => EmptyPartition(i))
+
+  override def compute(split: Partition, context: TaskContext): Iterator[InternalRow] = {
+    Iterator.empty
+  }
+}
+
+private[execution] case class EmptyPartition(index: Int) extends Partition
